backend/requirements.txt
groq
python-dotenv
requests
fastapi
pydantic
redis
pymongo 
yagmail
reportlab